pyproject.toml
[project]
name = "flight-delay-mlops"
version = "0.1.0"
description = "ВКР: ML-процесс прогнозирования задержек авиарейсов и классификации их причин"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "Georgij" }]

dependencies = [
    # core data
    "pandas>=2.2,<3.0",
    "numpy>=1.26,<3.0",
    "polars>=1.0,<2.0",
    "pyarrow>=15.0",

    # ML
    "scikit-learn>=1.5,<2.0",
    "xgboost>=2.1,<3.0",
    "lightgbm>=4.5,<5.0",
    "catboost>=1.2,<2.0",
    "imbalanced-learn>=0.13,<1.0",
    "optuna>=4.0,<5.0",
    "shap>=0.46,<1.0",

    # MLOps
    "mlflow>=2.17,<3.0",
    "dvc>=3.55,<4.0",

    # API
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "pydantic>=2.9,<3.0",

    # monitoring
    "prometheus-client>=0.21,<1.0",
    "structlog>=24.4,<25.0",

    # config
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
    "pytest>=8.3,<9.0",
    "pytest-cov>=5.0,<7.0",
    "pre-commit>=4.0,<5.0",
    "ipykernel>=6.29,<7.0",
    "jupyter>=1.1,<2.0",
    "matplotlib>=3.9,<4.0",
    "seaborn>=0.13,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ============================ ruff ============================
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks", ".venv", "data", "models", "mlruns"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "N",   # pep8-naming
    "C4",  # flake8-comprehensions
    "RET", # flake8-return
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802", "N803"]

# ============================ mypy ============================
[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["notebooks/", "data/", "mlruns/", ".venv/"]

# ============================ pytest ============================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
filterwarnings = ["ignore::DeprecationWarning"]
